Refactor app.py
